Implemented io generation via genrule
